Center image with a different ratio aspect
This commit solves [this
issue](https://github.com/hteumeuleu/H8UoEmbed/issues/3). A
`H8UoEmbed--ratio-16-9` class is added on `H8UoEmbed` as well as a
`H8UoEmbed-img--ratio-4-3` class on `H8UoEmbed-img` if the video player
has a 16:9 ratio and the image has a 4:3 ratio.

Styles were then added to make it work, based on [this
technique](http://www.mademyday.de/css-height-equals-width-with-pure-css
.html).